Fix UI animation typewriter (#4)
## Changes

Because strings are immutable, calling `gsub` without storing the result
is not accomplishing anything. Additionally, there was no comment
explaining what the `gsub` was doing. I fixed these issues, and also
replaced `wait` with `task.wait` to adopt modern best practice.

## Checks

By submitting your pull request for review, you agree to the following:

- [X] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [X] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [X] To the best of my knowledge, all proposed changes are accurate.

---------

Co-authored-by: IgnisRBX <bsorrentino@roblox.com>