Bump the memory requirements to 4GB from 2GB (#45)
## Changes

<!-- Please summarize your changes. -->

##### First time attempting to contribute to Open Source, hope it goes
alright...

**Short Summary:** Studio has outgrown my existing 2GB computer, making
Studio unusable, with it being total opposite of a smooth experience.
The documentation still says it should be smooth with that low memory,
although recent changes have made it anything but that. So, bump up the
memory requirements.

**Long Summary:**
This won't be taken lightly, so here's how it is in detail. My main
computer is dated 2009/2010, low-end of it's time, with the following
specifications:
- LGA775 Pentium E6500 Dual-Core @ 2.93 GHz
- 2.00 GB of system memory, 1-channel, DDR3
- Intel G41 Express Chipset (DirectX 10)
- Windows 7 x64
- A 7200RPM HDD, ~80GB free
- 8 M**b**/s internet connection (1 Mb/s upload)

Looks like it matches almost exactly the minimum requirements. The
documentation says that Studio should run smoothly. How smoothly does it
run in this configuration? Well...
Calling it not smooth would be a understatement: it's, at best,
*glacier.*
- It takes 3 minutes to start up with a full minute and a half of it
being frozen [not responding] on the login screen.
- The home page alone takes over 780 MB, a simple baseplate takes over
1.2 GB and pressing Play on a simple baseplate takes 1.8 GB of memory
(commit). And each stage takes atleast a minute, with the disk
utilization pegged at 100% HDD usage.
- Anything more simple than a baseplate is borderline unusable, with
simplest games taking far longer to make. It also limits the amount of
developer tools you can use. Don't even dare to work with Rojo, because
Studio will time out when starting CoreScripts.
In 2019, 2020 and early 2021 it was completely usable. It took a sudden
turn after that, and I keep a copy of a version of Studio just before
the sudden turn.

While you could delete built-in plugins and scoop up a dozen FFlags by
your own to reduce memory usage to make it usable, it only gets usable
at simple games and you still have to get a friend to press Play and
report errors 24/7 in order to avoid the thrashing.
And yes, I am putting my **own** PC out of the minimum system
requirements.

<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->

There is a bug report that [complains about memory deficiencies in
plugins](https://devforum.roblox.com/t/unused-or-internal-built-in-plugins-are-always-loaded-causing-excess-memory-usage/2466797),
which was not answered by any Roblox-official triager and is assumed to
be not looked at.
Fixing this almost certainly requires communication across 63 teams (1
for each plugin). Since it's unlikely to be fixed, just make it clear
that Studio will be laggy by bumping the memory requirement to 4GB.
###### There are also further memory deficiencies in Studio, which the
most important one includes built-in plugins having their own Roact17
and package copies (which results in plugins being 142,521,624 bytes
uncompressed). With more and more built-in plugins coming every update,
it gets harder and harder to even open Studio without modifications.
###### Roblox Player, even after post-Byfron, is completely usable as
long as you avoid the desktop app. It's just Studio that is unusable
with low memory, so no need to bump *that* one.

## Checks

By submitting your pull request for review, you agree to the following:

- [X] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [X] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.