Update create-player-hazards.md (#179)
## Issue
The current script includes a fail-safe for cases where the
`Players.PlayerAdded` event might not work perfectly using a for-loop
for existing players. If a player collects coins before the
`Players.PlayerAdded` event is executed or before the 'for loop'
executes, their coin count would be reset to zero. This pull request
addresses this issue by introducing a check to prevent resetting the
count if the player already has coins.

## Changes
The code change involves modifying the `onPlayerAdded` function to check
if the player already has coins (`currentCoins`). If `currentCoins` is
`nil or zero`, it returns zero. This ensures that a player's coin count
is only reset if they don't have any coins collected.

Code Change Illustration:
```lua
-- Before
for _, player in ipairs(Players:GetPlayers()) do
	onPlayerAdded(player)
end
    -- Resets the coin count to zero without checking for current coins as the method 'onPlayerAdded' does not check for currentCoins

-- After
local function onPlayerAdded(player)
	-- Check if the player already has coins
	local currentCoins = PlayerData.getValue(player, COIN_KEY_NAME)
	if currentCoins == nil or currentCoins == 0 then
		-- Reset player coins to 0 if they don't have any coins
		updatePlayerCoins(player, function(_)
			return 0
		end)
	end
```

Since the 'for loop' invokes the onPlayerAdded method, the fix was
implemented there. Alternatively, it can be written in the for loop code
block.

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

---------

Co-authored-by: Andrew Etter <aetter@users.noreply.github.com>
Co-authored-by: IgnisRBX <43388550+IgnisRBX@users.noreply.github.com>