Update BrickColor.yaml (#337)
Example used RGB values 0-255, when BrickColor takes Color3 values 0-1
and constructs the nearest BrickColor.

## Changes
Switched the colors from an RGB byte to a decimal between 0-1 and
slightly modified the comment(wasn't sure how to word it).
<!-- Please summarize your changes. -->

<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->

![image](https://github.com/Roblox/creator-docs/assets/85803876/c902b38f-274a-4686-90af-6b99d2cc6df5)
As you can see, I am creating a new BrickColor with the values (255, 2,
2), which you would expect to be converted into a quite red color.
Instead, these are interpreted as 0-1 values. Because the closest
BrickColor to these values is Institutional white, that is the
BrickColor applied.


![image](https://github.com/Roblox/creator-docs/assets/85803876/2f53fb5f-a8b5-44fc-bc85-0970767b4703)
Whereas if I divide all the values by 255, I get the expected result.
## Checks

By submitting your pull request for review, you agree to the following:

- [X] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [X] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [X] To the best of my knowledge, all proposed changes are accurate.

---------

Co-authored-by: IgnisRBX <43388550+IgnisRBX@users.noreply.github.com>