Correct description of global unpack's parameters (#341)
Correct the description of the unpack parameters.

## Changes

The description of `unpack`'s `j` parameter was wrong. It is not a
count, it is the index of the last element to `unpack`. Updated the
description of the `i` parameter to match the new wording of the `j`
parameter.

## Checks

By submitting your pull request for review, you agree to the following:

- [X] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [X] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [X] To the best of my knowledge, all proposed changes are accurate.

---------