Update require RobloxGlobals.yaml (#361)
## Changes

<!-- Please summarize your changes. -->
The description for "require" was very short and should provide more
context on its behavior, especially regarding one of the major uses of
modulescripts, that is to have a method of sharing values across
scripts. Also does not elaborate on client server boundary affecting
require
<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.