Update LuaGlobals.yaml getfenv stack level example
The getfenv by stack level example was completely incorrect. It was attempting to access a locally defined variable, overshadowed it's own password variable in the whatIsThePassword function, and used the wrong level for retrieving the stack of the function which called whatIsThePassword. This new example uses a ModuleScript to show how you can retrieve global variables from other scripts that you are not able to normally.