Clarify explosion behavior (#462)
## Changes
Explosion summary claimed that explosions are automatically `destroyed`
after a short time. However, terminology is important here, as in
practice explosions are only `removed`, so their parent is not locked
and the connections to the explosion are not disconnected. By reading
`destroyed`, readers may fail to understand this nuance, leading to them
creating memory leaks due to not cleaning up their connections.

Related devforum topic:
https://devforum.roblox.com/t/code-sample-for-explosionhit-has-a-memory-leak/2873076/6

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

---------

Co-authored-by: IgnisRBX <43388550+IgnisRBX@users.noreply.github.com>