Don't mark `Instance.DescendantRemoving` as deprecated (#481)
## Changes

At some point in the past, `Instance.DescendantRemoving` was labeled as
deprecated. Given it is not actually deprecated, this was clearly a
mistake and is misleading. This PR marks it as not deprecated.

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.