DMCAGuidelinesUpdate (#660)
## Changes

<!-- Please summarize your changes. -->

- Corrected the location of more information about the Roblox DMCA
Process

Previously, users were directed to Section 5 of the Roblox Terms of
Service. However, the fifth section is actually about payments and
refunds. I corrected it to Section 6, which is about Intellectual
Property and UGC that covers these topics.

<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.