Removed Historical Functions
GetCore no longer supports these functions. They are not deprecated, because functionality is not only unsupported, but completely gone. Querying any of these will return nil every time and never anything else. This has been the functionality for multiple years.

ChatActive still works as expected, so I left it on the page.