Update GuiButton.yaml (#686)
## Changes

Connecting to the `GuiButton.Activated` event on the server works, but
the script never receives any signals. Whereas, doing the same thing on
the client, the script receives signals. This pull request clarifies
which type of script to use for the event.

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.