Fix Restaurant Soup Simulator Example (#712)
## Changes

<!-- Please summarize your changes. -->

<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->

This pull request updates the Restaurant Soup Simulator example on the
pairs and ipairs docs page. The `isInSoup` table was renamed to
`selectedIngredients` because the name `isInSoup` didn't make sense with
how it was used, and the conditional at the end was fixed to check the
length of the table (previously it always evaluated to true because the
table did exist, but now it only evaluates to true when the table has at
least one element).

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.