deprecation-message-additions (#715)
## Changes

<!-- Please summarize your changes. -->

I added some deprecation messages to the "Roblox Globals" Documentation
for some methods that have been replaced. This is to add further context
as to why they are replaced, and to further direct readers to the newer
methods.

<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->

I have none, its mostly for convinience to the reader.

## Checks

By submitting your pull request for review, you agree to the following:

- [X] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [X] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [X] To the best of my knowledge, all proposed changes are accurate.

---------