Update TeleportService.yaml
https://devforum.roblox.com/t/teleportasync-not-returning-a-teleportasyncresult/3145215/13

See here. The API documentation here acts as if TeleportAsyncResult will always be returned and that TeleportOptions must be passed. It does not notify of the result that users may experience if TeleportOptions is not passed. As such, it should be denoted as an optional parameter and the result marked as an optional result.