Update TeleportService.yaml (#805)
## Changes

<!-- Please summarize your changes. -->
The API documentation here acts as if TeleportAsyncResult will always be
returned. It does not notify of the result that users may experience if
TeleportOptions is not passed. As such, it should be denoted as an
optional parameter and the result marked as an optional result.

<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->
See:
https://devforum.roblox.com/t/teleportasync-not-returning-a-teleportasyncresult/3145215/13

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

---------