Add Secrets to game-settings.md (#845)
## Changes

<!-- Please summarize your changes. -->

Secrets can now be created and configured via the game settings security
menu for local studio sessions. This requires that HTTP requests be
allowed.

<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->


https://devforum.roblox.com/t/secrets-store-general-availability/2892154/72

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

---------