Fix hotkey for swapping between Move and Rotate moves in AnimationEditor doc (#975)
## Changes

Fixes out dated entry in Animation editor. The hotkey for swapping
between Move and Rotate has changed from `ctrl+2` / `ctrl+4` to just
`R`. This devforum post confirms it (works for me locally as well):
https://devforum.roblox.com/t/how-do-you-switch-to-the-movement-mode-in-the-animation-editor/1791442

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

---------