Remove use of the word "force" in describing ApplyImpulse methods
Since the various ApplyImpulse functions apply a "pure" impulse (resulting in an instantaneous change in velocity), using the word "force" is misleading. A force must be applied over some finite time span to produce a velocity change, whereas an impulse is applied instantaneously. This has led to confusion among developers.