Update HttpService.yaml with HTTP methods, explanations, and resources.  (#978)
## Changes

Updated HttpService.yaml to provide a short description of each of the
HTTP methods, if they are permitted via Roblox scripts, and Mozilla
resources for each. This is a duplicate of pull request
https://github.com/Roblox/creator-docs/pull/966 as the checks did not
seem to recognize the changes made to
tools/checks/utils/allowedHttpLinks.txt to allow the links to Mozilla
resources on each of the HTTP methods.

The only method not currently permitted by Roblox is the CONNECT method
which creates a secure information tunnel between the requested and
originating server. Makes sense why this isn't allowed, but its useful
information to include nonetheless.

---------