Update RaycastResult.yaml (#1015)
## Changes

<!-- Please summarize your changes. -->
Changed the references from "instance" to BasePart. Referring to the
type of the intersected object as "instance" is too generic and may be
confusing, since Models and Folders are also instances that some
intersected object could have been a child of, but raycasts can only
return a reference to the object in of itself and never the parent
model. The (poorly named) _instance_ property of RaycastResult can only
ever hold a basepart or a terrain cell, and terrain descends from
basepart so basepart is accurate and more specific than instance.


<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->

## Checks

By submitting your pull request for review, you agree to the following:

- [x ] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x ] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x ] To the best of my knowledge, all proposed changes are accurate.

---------