Fix typo "Spacial" to "Spatial" in tag-table.md (#1041)
## Changes

The tag as shown in MicroProfiler is "frameStep/SpatialFilter" not
"frameStep/SpacialFilter".

I fixed the spelling.

## Checks

By submitting your pull request for review, you agree to the following:

- [X] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [X] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [X] To the best of my knowledge, all proposed changes are accurate.