Clarity around the sender/receiver in ShouldDeliverCallback. (#1050)
## Changes

<!-- Please summarize your changes. -->

Elaborated on the sender/receiver structure of
`TextChannel.ShouldDeliverCallback`. My co-worker got confused while he
was working on a problem with us and thought the callback only gave you
information about who sent the message, since in Play Solo you would be
the only sender and recipient of the message. After confirming how I
suspected it actually worked, I went in and made this documentation
change so others hopefully don't have similar confusion in the future.

<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->

## Checks

By submitting your pull request for review, you agree to the following:

- [X] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [X] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [X] To the best of my knowledge, all proposed changes are accurate.

---------