Fix intro-to-module-scripts.md having a mixup (#1067)
## Changes

<!-- Please summarize your changes. -->
to be specific, it says to create a ModuleScript in ServerStorage, but
it is previously stated that they are commonly placed in
ServerScriptService, so this fixes it
<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->
No links are needed, as the file and image right below the line I
changed both state that it should be in ServerScriptService
## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [X] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [X] To the best of my knowledge, all proposed changes are accurate.