Update GlobalDataStore.yaml (#1069)
Updated to create new versions using SetAsync, but only once every hour,
with at least an hour between each version

## Changes

<!-- Please summarize your changes. -->
Fixed the Misleading Info On The Doc , and updated it according to this
Post From Dev_Forum


https://devforum.roblox.com/t/upcoming-changes-to-data-stores-versioning/3042258/1


<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->

https://devforum.roblox.com/t/upcoming-changes-to-data-stores-versioning/3042258/1

## Checks

By submitting your pull request for review, you agree to the following:

- [✔] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [✔ ] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [✔ ] To the best of my knowledge, all proposed changes are accurate.

---------