Add note that external tools are not managed by Roblox (#1071)
## Changes

This guide makes a reference to several `vscode` extensions, it may be
useful to mention that these tools are not maintained by Roblox, but
instead, the community.

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

---------