Update LineJoinMode.yaml (#1074)
Added figures from the UIStroke class page, as images are referenced in
the table although none exist Changed reference to the "Applying Strokes
Article" to reference the UIStroke class page (said article does not
exist).

This documentation error was brought to my attention by
[@XAXA](https://devforum.roblox.com/u/xaxa) in [this devforum
post](https://devforum.roblox.com/t/enumlinejoinmode-documentation-appears-to-be-outdated/3603081).

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

---------