Update Animator.yaml (#1078)
## Changes

<!-- Please summarize your changes. -->
put a note under LoadAnimation documenting possible hard error

<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->
I ran into this cryptic error "Cannot load the AnimationClipProvider
Service." when trying to load an animation. I did some searching and the
root cause seems to be calling load animation on an animator that isnt
yet parented to the workspace. I have no clue why this is a necessity at
all, let alone pressing enough to warrant a hard error. This seems to be
a long standing problem that occurs on occasion for humanoids and the
regular player model as well, causing some users to have broken
animations from the default animation script erroring. See this thread

https://devforum.roblox.com/t/error-cannot-load-the-animationclipprovider-service/1639315/98
## Checks

By submitting your pull request for review, you agree to the following:

- [x ] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [ x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [ x] To the best of my knowledge, all proposed changes are accurate.

---------

Co-authored-by: IgnisRBX <43388550+IgnisRBX@users.noreply.github.com>