Update webhook-notifications.md (#1066)
## Changes

<!-- Please summarize your changes. -->

<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->
I added missing webhook notification types, fixed the formatting of json
and fixed the javascript example because the "replay attack prevention"
did not work at all. If there was no signature sent in the header,
validation was skipped. And if the signature was present but it was
incorrect, the function did not return.
## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

Co-authored-by: IgnisRBX <43388550+IgnisRBX@users.noreply.github.com>