Update table docs (#1092)
## Changes

* Removes reccomendation of `pairs()` and replaces it with generic
iteration
* Renames the deepCopy function to deepClone
* Makes the deepClone function use table.clone
* Fix deepFreeze erroring if a value is already frozen
* Explains dot indexing for dictionaries `.key`
* Removes warning for using ipairs to iterate over a dictionary, as
ipairs is unnessary to use unless wanting to stop at the first nil in a
array with holes (a behavior that is very rarely desired)

The last one is a bit iffy so feel free to reverse it. Also unsure if
the wording for `Write to dictionaries` is the best, but I think the
added warning clears it up.

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

---------