Document InputObject behavior in CAS action functions (#1110)
## Changes

- Document how the InputObject passed to a ContextActionService action
function won't correspond to a bound input when the function is called
with Cancel state.
- Since every entry in the table of function parameters would have an
extra detail/note about it, it seemed unnecessary to keep those notes
outside of the table, so they have been put in to the table.

## Checks

By submitting your pull request for review, you agree to the following:

- [X] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [X] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [X] To the best of my knowledge, all proposed changes are accurate.

---------

Co-authored-by: IgnisRBX <43388550+IgnisRBX@users.noreply.github.com>