Relocate ProcessReceipt API notes
Update document for readability and improved organization. 

Moved the ProcessReceipt API comments from below the header of "Sell your developer product" to it's appropriate location, "Outside your experience." This is to avoid misleading developers into believing that ProcessReceipt is required for all transactions, when it is actually only required for transactions outside of the experience. Also improved some verbiage for better readability.