Removed duplicates of debug.info() (#1085)
## Changes

<!-- Please summarize your changes. -->

<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->
I ran into the debug.info() thing 3 times while scrolling. This should
fix it. One of the duplicates had a different code sample that the other
two, so I cut and pasted it into the one debug.info() that I made.

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

---------

Co-authored-by: IgnisRBX <43388550+IgnisRBX@users.noreply.github.com>