Add example usage function for pcall() (#1114)
## Changes

The documentation can be somewhat confusing when trying to learn how to
correctly implement pcalls as pointed out by @junveld in [this forum
post](https://devforum.roblox.com/t/feedback-on-luau-globals/3664364). I
have added a small piece of example code for how to use pcalls to do
function calls that even when returning an error, will allow the rest of
the script to continue as normal.

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

---------