Update Player CharacterAppearanceLoaded (#1134)
When using custom character implementations such as using a
StarterCharacter, people should handle the character appearance
themselves. This serves as a note, a warning that
CharacterAppearanceLoaded does not fire for custom characters.

## Changes

<!-- Please summarize your changes. -->
Added an additional note for when this event
(`CharacterAppearanceLoaded`) should not be used (such as using
StarterCharacters), because it will not fire and I think this note will
save a lot of people from headache debugging.
```
      For custom character implementations, such as using a `Class.StarterPlayer|StarterCharacter`,
      Use `Class.Player.CharacterAdded` and handle your own accessories.
```

<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->
Reference: 

https://devforum.roblox.com/t/playercharacterappearanceloaded-does-not-fire-if-there-is-a-startercharacter-present/3684116

## Checks

By submitting your pull request for review, you agree to the following:

- [✅] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [✅ ] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [✅ ] To the best of my knowledge, all proposed changes are accurate.

---------

Co-authored-by: IgnisRBX <43388550+IgnisRBX@users.noreply.github.com>