Update AnimationTrack.yaml (#1141)
Removing a round off error.

## Changes

<!-- Please summarize your changes. -->
Deleteing a ton of zeros and an extra one. No idea how it would benefit
anyone as it's a default value. The documentation also uses 0.1 in the
first paragraph of the Play() method.
<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [X] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [X] To the best of my knowledge, all proposed changes are accurate.

---------

Co-authored-by: IgnisRBX <43388550+IgnisRBX@users.noreply.github.com>