Update documentation for GetWaveformAsync (#1158)
## Changes

These are both live but it looks like the documentation was never filled
in for them.

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

---------

Co-authored-by: IgnisRBX <43388550+IgnisRBX@users.noreply.github.com>
Co-authored-by: rbxphogen <89100144+rbxphogen@users.noreply.github.com>