Update meshes.md (#1171)
## Changes

Added alt="" to some images to ensure accessibility or if the user has
slow internet.

Changed wording from MaterialVariant being a child of the mesh part,
which is totally incorrect, MaterialVariant's only work in
MaterialService and must be set in the mesh's MaterialVariant property
otherwise it does not work.

Changed wording how to add a SurfaceApperance or MaterialVariant to
reflect how it works ^.

If you require more information about these changes please leave a
comment and I'll be happy to awnser any questions.

N/A

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

---------

Co-authored-by: IgnisRBX <43388550+IgnisRBX@users.noreply.github.com>